changedTestMuseumAPI.py
Added two more search queries: type of work and general search.

Added 'imgonly=True' to mandatory search terms (line11). WIthout it, the search results that do not have an image, return None and listOrUrls becomes unappendable. As a result, the program cannot download any of the images even if they were returned in the search results.
Imgonly=True fixes this.